+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+- Slack file upload requests use form encoding for `files.getUploadURLExternal` and multipart upload to the presigned URL (fixes `invalid_arguments` after v2 migration).
+
 ## [0.4.2] - 2026-05-30
 
 ### Fixed